Use our own implementation of NavigationView
This allow to expose summary and number notification of items
and display them if required in the navigation view.

Since can't create a wrapper around MenuItemImpl without rewrite the entire
support implementation, just use the title to fill the other values of
interest (summary and number notifications). Just use a separator "|" from
title string to determine <title>|<summary>|<notification count>
Callers must guarantee the sanity of the title string. This is a bad hack :(,
but allow us to display the information we want on screen.

Signed-off-by: Jorge Ruesga <jorge@ruesga.com>