Correct text hint when reverting/restoring a change
Currently, when reverting or restoring a change, the EditText
in the dialog hints to type a "Commit Message".

However, there is no commit message to be entered. Instead,
a comment is entered, just like a normal comment on any change.

Hence, correct the text hint to display "Comment on this change".