Reduce avatar image fetching pressure
Since we returned different avatar urls every time,
we try to fetch all the urls every time until
it finds a valid one, which made Picasso to fail to
load the proper cached imaged and refetch all the images
every time. Since we have the url cached in memory, just
serves the original array to remove all the invalid entries
when a valid url was fetched, so the valid cached image is
used always as a resource.

Signed-off-by: Jorge Ruesga <jorge@ruesga.com>