Some repositories seems to force accounts/{account-id} endpoint with a trailing /
Rview is unable to login with AOSPA repository because this is enforcing a trailing /
for accounts/{account-id} endpoint (even documentation doesn't define this trailing / and
other repositories doesn't present such restriction), which fail with 403 error.
Since other repositories accept the trailing / in this endpoind, just update it
to bypass AOSPA restriction.

Signed-off-by: Jorge Ruesga <jorge@ruesga.com>